refactor: use new createAccountFromImplicit thunk instead of action (#2379)
