+++ packages/frontend/src/components/accounts/create/initial_deposit/InitialDepositWrapper.js
@@ -7,8 +7,9 @@ import { useSelector, useDispatch } from 'react-redux';
 
 import { MIN_BALANCE_TO_CREATE } from '../../../../config';
 import { Mixpanel } from '../../../../mixpanel';
-import { createAccountFromImplicit, redirectTo, checkIsNew } from '../../../../redux/actions/account';
+import { redirectTo, checkIsNew } from '../../../../redux/actions/account';
 import { showCustomAlert } from '../../../../redux/actions/status';
+import { createAccountFromImplicit } from '../../../../redux/slices/account/createAccountThunks';
 import { actions as createFromImplicitActions } from '../../../../redux/slices/createFromImplicit';
 import { actions as flowLimitationActions } from '../../../../redux/slices/flowLimitation';
 import { getSignedUrl } from '../../../../utils/moonpay';
@@ -99,7 +100,7 @@ export function InitialDepositWrapper({ history }) {
         await Mixpanel.withTracking("CA Create account from implicit",
             async () => {
                 setClaimingAccount(true);
-                await dispatch(createAccountFromImplicit(accountId, implicitAccountId, recoveryMethod));
+                await dispatch(createAccountFromImplicit({ accountId, implicitAccountId, recoveryMethod }));
             },
             async (e) => {
                 console.warn(e);
